refactor: productList 상태 있는 것과 없는 것으로 나누기
